Refactor : 현재 유저정보 App 컴포넌트에서 가져와서 Context로 전달
