+++ src/hooks/useAccessToken.ts
@@ -1,0 +1,14 @@
+import { useEffect, useState } from "react";
+
+export default function useAccessToken() {
+  const [accessToken, setAccessToken] = useState<string | null>(null);
+
+  useEffect(() => {
+    const token = localStorage.getItem("accessToken");
+    if (token) {
+      setAccessToken(token);
+    }
+  }, []);
+
+  return accessToken;
+}
