+++ tsconfig.json
@@ -15,6 +15,7 @@
     "incremental": true,
     "baseUrl": "./",
     "paths": {
+      "@pages/*": ["pages/*"],
       "@components/*": ["pages/components/*"],
       "@styles/*": ["pages/styles/*"],
       "@apis/*": ["src/apis/*"],
